Transplanting the camera system into a tools module and syncing it with the frontend.
