--- app/static/webasset/css/base.css
@@ -1,61 +1,0 @@
-.header {
-  overflow: hidden;
-  background-color: rgba(0,0,0,0);
-}
-
-/* Style the header links */
-.header a {
-  float: left;
-  color: black;
-  text-align: center;
-  padding: 12px;
-  text-decoration: none;
-  font-size: 18px;
-  line-height: 25px;
-  border-radius: 4px;
-}
-
-/* Style the logo link (notice that we set the same value of line-height and font-size to prevent the header to increase when the font gets bigger */
-.header a.logo {
-  font-size: 25px;
-  font-weight: bold;
-}
-
-/* Change the background color on mouse-over */
-.header a:hover {
-  background-color: rgba(0, 0, 0, 0);
-  color: black;
-}
-
-/* Style the active/current link*/
-.header a.active {
-  background-color: rgba(30, 229, 255, 0.8);
-  color: white;
-}
-
-/* Float the link section to the right */
-.header-right {
-  float: right;
-}
-.header-center {
-  float: right;
-}
-.header-left {
-  float: right;
-}
-/* Add media queries for responsiveness - when the screen is 500px wide or less, stack the links on top of each other */
-@media screen and (max-width: 500px) {
-  .header a {
-    float: none;
-    display: block;
-    text-align: left;
-  }
-  .header-right {
-    float: none;
-  }
-}
-body{
-        font-family:Arial;
-    font-size: 20px;
-}
-
