+++ app/templates/frameUnit/_video.html
@@ -1,0 +1,1 @@
+<img class="videosrc" src="{{ url_for('video_feed') }}">
