+++ database/see.json
@@ -1,1 +1,1 @@
-{"sys_v": 0.5, "default_app": "simplestreamer", "counter_clockwise_rotation_count": 0}
+{"sys_v": 0.5, "default_app": "", "counter_clockwise_rotation_count": 2}
